Concentrate model implementation differences to hantekdsocontrol
* Remove usbdevice special case for the 6022 model and introduce a feature flag instead.
* Remove all specific model checks across the hantekdsocontrol class and introduce feature flags instead.

Bigger picture: All model differences should be described in the controlsettings and specification structures
within the constructor of hantekdsocontrol. The rest of the class should be model independant code or marked as
very model dependant code. This should allow to split the class into smaller parts to realize Separation of concerns.

TODO: getSamples() has basically two code paths: One for the 6022Bx and one for all other devices.
I suggest getSamples() and getSamples6022Bx() for instance.