Fixed crashing of application on FreeBSD
Currently the application was crashing on FreeBSD due to the device
list being released (incorrectly) even though reference counts
are correctly incremented. This seems to be a bug specific to the
libusb implementation on FreeBSD. The solution is currently to
disable the release of the device list after enumeration which would
introduce a potential resource leak in case the selection dialog is
launched multiple times.

Signed-off-by: tspspi <11219971+tspspi@users.noreply.github.com>