new image/pdf export and print
The solution takes either a screenshot of the mainwindow as it is or from the
dsowidget only with the option to change the colors from screen to printer.
A zoomed scope screen can be captured in double or (optional) original hight.
Storage is possible as png/jpg/bmp or pdf.

Signed-off-by: Martin <Ho-Ro@users.noreply.github.com>